# specs/001-clinic-scheduling/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Clinic Patient Scheduling API
  version: 0.1.0
  description: >
    Tenant-scoped REST API for clinic patient scheduling. Authentication is via a
    server-side session cookie that carries the user's center scope. Every non-super-admin
    endpoint is scoped to the caller's center and fails closed (403, no data) if scope
    cannot be established (Constitution Principle II). All times are UTC (RFC 3339).
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: double_booking }
        message: { type: string }
    Role:
      type: string
      enum: [super_admin, center_admin, doctor, reception]
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/Role' }
        email: { type: string, format: email }
        display_name: { type: string }
        is_active: { type: boolean }
        is_admin: { type: boolean }
        must_change_password: { type: boolean }
    UserCreate:
      type: object
      required: [role, email, display_name, temp_password]
      properties:
        role: { type: string, enum: [doctor, reception] }
        email: { type: string, format: email }
        display_name: { type: string }
        temp_password: { type: string }
        specialty: { type: string, nullable: true }
    Patient:
      type: object
      properties:
        id: { type: string, format: uuid }
        full_name: { type: string }
        clinic_identifier: { type: string }
        phone: { type: string, nullable: true }
        date_of_birth: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
    PatientCreate:
      type: object
      required: [full_name]
      properties:
        full_name: { type: string }
        clinic_identifier: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        date_of_birth: { type: string, format: date, nullable: true }
        notes: { type: string, nullable: true }
        confirm_possible_duplicate: { type: boolean, default: false }
    DuplicateWarning:
      type: object
      properties:
        code: { type: string, enum: [possible_duplicate] }
        candidates:
          type: array
          items: { $ref: '#/components/schemas/Patient' }
...
    AppointmentRequest:
      type: object
      properties:
        id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        reason: { type: string, nullable: true }
        preferred_from: { type: string, format: date, nullable: true }
        preferred_to: { type: string, format: date, nullable: true }
        urgency: { type: string, enum: [routine, soon, urgent] }
        expected_duration_minutes: { type: integer }
        notes: { type: string, nullable: true }
        status: { type: string, enum: [pending, fulfilled, declined] }
        decline_reason: { type: string, nullable: true }
        is_overdue: { type: boolean }
        resulting_appointment_id: { type: string, format: uuid, nullable: true }
    AppointmentRequestCreate:
      type: object
      required: [patient_id, urgency, expected_duration_minutes]
      properties:
        patient_id: { type: string, format: uuid }
        reason: { type: string, nullable: true }
        preferred_from: { type: string, format: date, nullable: true }
        preferred_to: { type: string, format: date, nullable: true }
        urgency: { type: string, enum: [routine, soon, urgent] }
        expected_duration_minutes: { type: integer }
        notes: { type: string, nullable: true }
    RequestDecline:
      type: object
      required: [decline_reason]
      properties:
        decline_reason: { type: string }
    Appointment:
      type: object
      properties:
        id: { type: string, format: uuid }
        doctor_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        clinic_id: { type: string, format: uuid, nullable: true }
        starts_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        status: { type: string, enum: [scheduled, completed, cancelled, no_show] }
        origin: { type: string, enum: [request, direct] }
        source_request_id: { type: string, format: uuid, nullable: true }
    AppointmentCreate:
      type: object
      description: >-
        Provide either patient_id (existing patient) or patient_name +
        patient_clinic_identifier (a walk-in, resolved/created on the fly).
      required: [doctor_id, starts_at, duration_minutes]
      properties:
        doctor_id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid, nullable: true }
        patient_name: { type: string, nullable: true }
        patient_clinic_identifier: { type: string, nullable: true }
        clinic_id: { type: string, format: uuid, nullable: true }
        starts_at: { type: string, format: date-time }
        duration_minutes: { type: integer }
        source_request_id: { type: string, format: uuid, nullable: true }
        confirm_patient_conflict: { type: boolean, default: false }
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        payload: { type: object }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }
  responses:
    Forbidden:
      description: Not permitted / out of center scope (default-deny)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Integrity conflict (e.g., double_booking, slot_taken, invalid_transition)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unprocessable:
      description: Validation error (e.g., outside_availability, off_grid, patient_conflict)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
security:
  - sessionCookie: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: List staff in caller's center (center-admin)
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [users]
      summary: Create a doctor or reception account in caller's center (center-admin)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserCreate' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /users/{userId}:
    put:
      tags: [users]
      summary: Edit a staff account (center-admin)
      parameters: [ { name: userId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { display_name: { type: string }, is_active: { type: boolean }, is_admin: { type: boolean } } } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
...
  /patients:
    get:
      tags: [patients]
      summary: Search patients by name or phone within caller's center
      parameters: [ { name: q, in: query, schema: { type: string } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Patient' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [patients]
      summary: Register a patient (warns on possible duplicate unless confirmed)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PatientCreate' } } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Patient' } } } }
        '409': { description: Possible duplicate (resubmit with confirm_possible_duplicate), content: { application/json: { schema: { $ref: '#/components/schemas/DuplicateWarning' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---- availability ----
...
  /requests:
    get:
      tags: [requests]
      summary: List appointment requests (reception sees the pending queue; doctor sees own)
      parameters: [ { name: status, in: query, schema: { type: string, enum: [pending, fulfilled, declined] } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/AppointmentRequest' } } } } } }
    post:
      tags: [requests]
      summary: Doctor submits an appointment request
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentRequestCreate' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentRequest' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
  /requests/{requestId}/fulfill:
    post:
      tags: [requests]
      summary: Reception fulfills a request by booking a concrete slot (creates an Appointment)
      parameters: [ { name: requestId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentCreate' } } } }
      responses:
        '201': { description: Appointment created and request fulfilled, content: { application/json: { schema: { $ref: '#/components/schemas/Appointment' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /requests/{requestId}/decline:
    post:
      tags: [requests]
      summary: Reception declines a request with a reason (notifies the doctor)
      parameters: [ { name: requestId, in: path, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RequestDecline' } } } }
      responses: { '200': { description: Declined, content: { application/json: { schema: { $ref: '#/components/schemas/AppointmentRequest' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }

  # ---- appointments ----
...
  /notifications:
    get:
      tags: [notifications]
      summary: List the caller's in-app notifications
      parameters: [ { name: unread, in: query, schema: { type: boolean } } ]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Notification' } } } } } }
  /notifications/{notificationId}/read:
    post:
      tags: [notifications]
      summary: Mark a notification read
      parameters: [ { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Marked read } }